refactor(fs): split places into focused modules
