+++ src/fs/places/mod.rs
@@ -1,0 +1,8 @@
+mod devices;
+#[cfg(target_os = "linux")]
+mod linux;
+mod resolution;
+#[cfg(test)]
+mod tests;
+
+pub(crate) use self::resolution::{build_sidebar_rows, home_dir, trash_dir};
